[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -2,5 +2,5 @@
 using namespace std;
 
 int main() {
-    int a,b; cin >> a >> b; cout << (a+b)*(a-b);
+    long long a,b; cin >> a >> b; cout << (a+b)*(a-b);
 }
